Simplify renderer. Updated textures.
The draw_line and clear_screen methods were writing independently of the
wallet and keyboard controls. I wanted to have one simple IO.write call
that wrapped the various draw methods that were responsible for
populating their own character lists. This simplifies the code quite a
bit. I also updated the textures characters for the walls and players
and increased map spacing so it's not so claustrophobic.