Prevent the player from being able to move multiple cells at once
Reworked the Navigator module so that a moving flag is set each time the
player is in motion. This flag is then caught in a guard clause for
subsequent calls to the move/2 method. The flag will automatically be
reset by the renderer each time the player is rendered allowing the
player to move again. This effectively limits player movement to one
cell per render cycle.